+++ apiphany-core/src/main/java/org/apiphany/client/http/HttpClientFluentAdapter.java
@@ -1,7 +1,16 @@
 package org.apiphany.client.http;
 
+import java.util.List;
+import java.util.Map;
+
 import org.apiphany.ApiClientFluentAdapter;
+import org.apiphany.ParameterFunction;
+import org.apiphany.RequestParameters;
+import org.apiphany.http.HttpContentType;
+import org.apiphany.http.HttpHeader;
 import org.apiphany.http.HttpMethod;
+import org.apiphany.io.ChunkedBinary;
+import org.apiphany.io.ContentType;
 
 /**
  * Fluent adapter for HTTP calls.
@@ -129,5 +138,45 @@ public class HttpClientFluentAdapter {
 	 */
 	public ApiClientFluentAdapter method(final HttpMethod method) {
 		return apiClientFluentAdapter.method(httpExchangeClient.method(method));
+	}
+
+	/**
+	 * Sets the request body as URL-encoded form parameters. Automatically sets the {@code Content-Type} header to
+	 * {@code application/x-www-form-urlencoded}.
+	 *
+	 * @param params the form parameters
+	 * @return this
+	 */
+	public ApiClientFluentAdapter form(final Map<String, List<String>> params) {
+		var encodedParams = RequestParameters.encode(params, apiClientFluentAdapter.getCharset());
+		var body = RequestParameters.asString(encodedParams);
+		return apiClientFluentAdapter.body(body)
+				.header(HttpHeader.CONTENT_TYPE, ContentType.APPLICATION_FORM_URLENCODED);
+	}
+
+	/**
+	 * Sets the request body as URL-encoded form parameters. Automatically sets the {@code Content-Type} header to
+	 * {@code application/x-www-form-urlencoded}.
+	 *
+	 * @param paramFunctions the parameter functions
+	 * @return this
+	 */
+	public ApiClientFluentAdapter form(final ParameterFunction... paramFunctions) {
+		return form(RequestParameters.of(paramFunctions));
+	}
+
+	/**
+	 * Sets the request body as a multipart form-data body. Automatically sets the {@code Content-Type} header to
+	 * {@code multipart/form-data; boundary=...}.
+	 *
+	 * @param body the multipart body
+	 * @return this
+	 */
+	public ApiClientFluentAdapter multipart(final ChunkedBinary body) {
+		return apiClientFluentAdapter.body(body.toByteArray())
+				.header(HttpHeader.CONTENT_TYPE, HttpContentType.builder()
+						.contentType(ContentType.MULTIPART_FORM_DATA)
+						.boundary(body.getBoundary())
+						.build());
 	}
 }